backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "warehouse"
version = "0.1.0"
description = "Home warehouse management system"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "litestar[standard]>=2.0.0",
    "advanced-alchemy>=0.7.0",
    "asyncpg>=0.29.0",
    "msgspec>=0.20.0",
    "granian>=0.9.0",
    "passlib[argon2]>=1.7.4",
    "pyjwt>=2.9.0",
    "sniffio>=1.3.0",
    "rocketry>=2.5.0",
    "rq>=1.16.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "faker>=33.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/warehouse"]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src", "e2e"]

